agents/detection/requirements.txt
# Detection Agent specific requirements
google-cloud-bigquery>=3.13.0
google-cloud-monitoring>=2.16.0
google-cloud-logging>=3.8.0
google-cloud-pubsub>=2.18.0
google-cloud-firestore>=2.13.0
pyyaml>=6.0
python-dateutil>=2.8.2
psutil>=5.9.0
fastapi>=0.104.0
uvicorn>=0.24.0